Update patch set 2
Patch Set 2: Code-Review-1

Instead of duplicating templates it would be better to reuse global-jjb templates so that this project get's the benefits of centrally maintained jobs. We are working with projects to adopt Global JJB. This will reduce the maintenance burdon of all projects using global-jjb as it is maintained centrally and updates to the templates are easily passed to all projects using it.

It is able to support the use case here with sub-projects using our {project-name}-maven-jobs job-group which will pull in all of the verify job definitions.

Here is an example of one of the OpenDaylight projects using this job group:

https://github.com/opendaylight/releng-builder/blob/master/jjb/odlparent/odlparent.yaml

The case of sub-projects that can be manipulated with the existing "project-name" variable and in fact is the reason it exists. There are 2 ways of using this variable.

1. Including the top-level project:

 project-name: project-subproject

2. Only using hte subproject name:

 project-name: subproject

Whereas the "project:" definition is the Gerrit project to clone when building.

In order to pass in the path to the subproject's pom.xml this can be done using the "mvn-goals:" parameter which is also supported by global-jjb. So a subproject definition for this project might look something like:

 - project:
   name: cds-client
   project-name: cds-client
   project: common-dataservice
   mvn-goals: clean install -f /path/to/cds/client/pom.xml
   mvn-settings: cds-settings
   jobs:
     - '{project-name}-maven-jobs'
   stream:
     - master:
         branch: master

A tip: It is not necessary to wrap every string in yaml with single quotes. Instead only use it when your string includes a {variable} (curly braces) this will keep your YAML files looking more clean.

Hope this helps,
Thanh

Patch-set: 2
Label: Code-Review=-1