Update patch set 3
Patch Set 3:

Thanks Thanh, I've removed the templates-java.yaml file.

Please point me to a copy of the global templates.

Yes if possible I would like a bit more guidance to get started.  We have many simple project with a single POM, and some with multiple sub-parts (a client and a server).  Is there a global template for a project with multiple sub projects?

Patch-set: 3
Label: Code-Review=0