Update patch set 6
Patch Set 6: Code-Review-2

This is the wrong approach as it duplicates work done in global-jjb. I've instead resolved this in global-jjb by re-prioritizing the MAVEN_OPTIONS and MAVEN_PARAMS so that we can override the values correctly.

This is the global-jjb patch which resolves the issue. https://gerrit.linuxfoundation.org/infra/8176

I've also produced a patch release of global-jjb so that we can pull it in now and have merged an update patch in acumos' ciman repo via https://gerrit.acumos.org/r/286

This will allow you to add to the common-dataservice.yaml file a single line:

 mvn-params: '-Dorg.slf4j.simpleLogger.log.org.apache.maven.cli.transfer.Slf4jMavenTransferListener=info'

Which will override the default warn value correctly.

Hope this helps,
Thanh

Patch-set: 6
Label: Code-Review=-2