Update patch set 11
Patch Set 11:

> maybe it got deleted by an administrator?

Yes it seems to have been missing. I fixed that and it's passing that point now. The docker configuration seems to require some kind of credentials which I'm trying to reverse engineer and decrypt so I can make that part of the script pass on Sandbox.

Does Acumos have an IRC channel that we can better coordinate? If not feel free to ping me on the LF one in #lf-releng @ irc.freenode.net

Patch-set: 11