Update patch set 11
Patch Set 11:

Alright I think I've taken this as far as I can. I got the docker credentials over and the docker-login script is now failing with 404 not found. Unfortunately I'm not familiar with the docker setup so will need to defer this to part to someone who can.

Patch-set: 11
Label: Code-Review=0