Update patch set 17
Patch Set 17:

> Sure I can remove lf-provide-maven-settings; what about the
 > matching lf-provide-maven-settings-cleanup builder at the end?

Hi Chris,
Actually can we just reverse the call order of the lf-provide-maven-settings and the lf-infra-docker-login.  the docker login does a cleanup internally and is wiping out your maven settings which is why it cannot find the file.  This is a workaround.  Leave the last cleanup in place.

Patch-set: 17
Label: Code-Review=0