Add maven-install script.
Revise docker-enabled verify job to use goals clean install,
not clean deploy.  The global-jjb maven-build script uses goal
deploy, which causes the image to be deployed by verify step,
an undesirable side effect.  Add a shell script with goals clean
install so that the image is only built, not pushed.

Change-Id: I13007efc5e9c74c85512f5251ad3e1b9c2062dc7
Signed-off-by: Christopher Lott (cl778h) <clott@research.att.com>