Update patch set 1
Patch Set 1:

> The defaults.yaml already sets this variable to 30, don't see why
 > it's needed in this job.  Possibly, maybe, the python template
 > doesn't pass the build timeout value thru correctly?

It is how JJB the tool itself is designed. We have a different use case then the OpenStack project that originally designed JJB. Other orgs that also use JJB has raised this as an issue to the JJB project in the past. If you look at lines 48 - 55 of this Gerrit review [0] you can see that today as JJB the tool itself is designed the order of precedence for "defaults" or variable configuration overrides is as follow:

1) job-group definitions
2) project definitions
3) job-template
4) defaults.yaml

For LF projects we would prefer that defaults has a higher precedence then job-template but unfortunately that is not how it is currently designed and it is something we have to deal with. I'm in discussion with the OpenStack team upstream to see if we can allow custom precendence ordering but things like this will take time to resolve as this impacts all consumers of the JJB tool and not just us.

So what's this mean for us... If a variable is defined with a default value inside of a job-template that means you cannot use "defaults.yaml" to override that job-template value. You MUST use a project definition instead and is why this patch is necessary.

Let me know if this isn't clear.

[0] https://review.openstack.org/#/c/194497/3/specs/jenkins-job-builder_dsl-roadmap.rst

Patch-set: 1
Reviewer: Gerrit User 62 <62@655300a2-7a56-4504-9153-881edd907bea>
Label: Verified=0