Update patch set 2
Patch Set 2:

> Thanks for the link.  Should I just use the OpenO template
 > integration-csit.yaml ?

No that won't be needed....They are very nearly the same if not exactly the same template.  I only included that link so you could see how they work currently.  I thought it might help out.  As you can see they assume a certain project folder structure.  If that won't be a problem for the team building the tests then we can proceed.

Patch-set: 2