Update patch set 32
Patch Set 32:

> never mind I found the --draft flag for git review

Yep that's right. Although it's a feature that Gerrit upstream is removing in a future version. The problem with DRAFTs has always been it's confusing for people. If I recall they are going to introduce a new button instead that you can click to make a patch as not ready for reviews.

Patch-set: 32
Label: Code-Review=0