Update patch set 2
Patch Set 2:

Thank you for additional analysis.  If the LF feels it's the Acumos project responsibility to construct these templates entirely on our own, please say so.  I've been hoping since Feb 1 that a JJB expert would jump in and solve this problem in the LF recommended way, but 22 days of delay says that's probably not going to happen.

Patch-set: 2