Update patch set 3
Patch Set 3:

> What is the difference between "basebuild" and "builder"?  It all
 > kind of looks the same from where I'm sitting.  The vast majority
 > of our projects use the docker-enabled build minion; the rest use a
 > python-equipped minion.  What does "builder" provide?

Basebuild is deprecated, it's been migrated into "builder" for our packer images. And "docker" is "basebuild + docker" so very similar to "builder". In theory this patch shouldn't affect anyone using basebuild. What we need to be careful of is the 4c-4g part (CPUs & RAM) and make sure that no projects are getting the right size system for their build needs.

Patch-set: 3