Update patch set 1
Patch Set 1:

If you are copying the RTD merge job. It does not build anything. RTD has their own build servers for generating docs so the RTD merge job that is in global-jjb is simply a curl script that triggers an RTD builder over at readthedocs.io to run a build and deploy to RTD's own service.

I'm not sure what this rsync is doing but if it needs to rsync docs somewhere then there needs to be a step that runs `tox -e docs` so that it generates the docs first and then rsync can sync the files over.

Patch-set: 1
Label: Verified=0