Update patch set 60
Patch Set 60: Code-Review-1

Thank you for putting the new templates in acumos-jjb.  I suggest also moving the new macro file and the new shell scripts to the directory acumos-jjb (e.g., subdirectory shell for the shell scripts), in the interest of keeping custom-to-project things together, and revising the paths in the new templates suitably.

Patch-set: 60
Label: Code-Review=-1