Update patch set 62
Patch Set 62:

> I have these arguments with my team constantly.  Hurry up and merge
 > the code, they say, we'll document it tomorrow.  Somehow that doc
 > just never arrives.

To merge this thread and the other email thread, I do think it would be good to have a minimal amount of documentation in this initial patch. It could be as simple as adding a Python packaging section to the root ci-management README.md file.

Copying from the other thread, it would be nice to have:

#1) A small section that describes what a new Python project needs to minimally do in order to benefit from your work (i.e. changes to their jbb file).

#2) A small section that describes how pushed Python packages can be accessed, including the URLs. If the website URL and pip --extra-index-url URL are different, then they should both be provided.

Patch-set: 62