Update patch set 62
Patch Set 62:

I'm with Thanh.  Merge means, merge to master and rebuild.  It does not mean the result is ready for the world to use.  The project lead decides when a build is ready for the world by creating a release candidate, usually by posting "build release" in a gerrit review.  That's the established process in the LF RelEng world, for better or worse, so let's stick with it in the Pypi setup.

Patch-set: 62