Update patch set 62
Patch Set 62:

> > I'm with Thanh.  Merge means, merge to master and rebuild.  It
 > does
 > > not mean the result is ready for the world to use.  The project
 > > lead decides when a build is ready for the world by creating a
 > > release candidate, usually by posting "build release" in a gerrit
 > > review.  That's the established process in the LF RelEng world,
 > for
 > > better or worse, so let's stick with it in the Pypi setup.
 > 
 > I suppose this sounds like an odd workflow to me because I've been
 > following http://nvie.com/posts/a-successful-git-branching-model/
 > for several years. What you describe sounds like the role of the
 > develop/integration and release branches.
 > 
 > If this is a LF thing, then I don't want to diverge from their best
 > practices, and I'll play along.

Hi Paul,

The nvie workflow is model we are using, in this model you would have maintenance branches for 1.x, 2.x, 3.x, etc.... It is not a model that's suggesting you release on every patch merge. Someone's still making a decision of when to cut off an official release (by tagging) and then publishing a tagged build.

Hope this helps,
Thanh

Patch-set: 62