Update patch set 62
Patch Set 62:

> In the nvie link, the master branch is not written to directly but
 > through bug fixes and release branches. So it sounds like here the
 > master branch has been given the role of the develop branch.
 > Perhaps that's where my confusing was stemming from. If that's the
 > case, then I agree it makes sense to have a manual release event.

Yes, that's right. The nvie link describes a git branching model. It's not describing a release model (although they are very closely tied). In our case "master" is "develop" and stable branches like 1.x, 2.x, etc... are maintenance branches used to develop maintenance releases. We use release tags as one of the release vehicles. In my opinion there is a few things that should happen before a release.

1. Staged release, a job pushes to a staging area
2. Quality Control, a tester tests the staged release and confirms no major regressions
3. Project approval for release

For 1, usually a regularly scheduled job that's performing release ready builds (version numbers are production ready, packaged for distribution, etc...). This is typically the stage job.

At least in the OpenDaylight project these 3 steps are necessary and usually the TSC approves the release going out. Smaller projects that are not participating in the Simultaneous Release have more leeway and usually do their own approval (this might be the case for your project from the sounds of things).

Then performing the actual release, 2 things need to happen.

1. Git tag reproduced against the exact commit hash that Jenkins used to build the staged release. Usually performed by the PTL of the project and additionally git signed with their GPG key.

This is to ensure that if someone else wants to build your software from source they can reproduce from the same commit, and releasing the source code is a requirement for LF projects as well.

2. Artifacts pushed to release vehicle, pypi, release repo in Nexus, etc...

Patch-set: 62