Update patch set 7
Patch Set 7:

> Thanks for the assist here Jeremy.  Do you have any idea about this
 > baffling invocation exception?

Hi Chris, The last failure showed a ShellCheck error because it is not able to source a file that is determined at runtime through a variable.  I set a tag on it to ignore that spot, lets see if it works.  https://github.com/koalaman/shellcheck/wiki/SC1090
Jeremy

Patch-set: 7