Update patch set 7
Patch Set 7:

the repository for CSIT is called test-automation;  I just became aware of this patch for ci-management today. Can we have a meeting to bring me on board to this effort? It looks like the structure in test-automation might have to change due to the triggers defined.   
triggers:
      - gerrit-trigger-patch-submitted:
          server: '{server-name}'
          branch: '{branch}'
          files: 'test/csit/**/{project-name}/**'

Patch-set: 7