Update patch set 3
Patch Set 3:

> Jeremy I see you have added the appropriate YAML and now a job
 > "documentation-rtd-merge-master" appears in Jenkins.  It runs
 > successfully this CURL command:
 > 
 > curl -X POST --data version_slug=latest https://readthedocs.org/build/acumos-documentation
 > 
 > Forgive the obvious question, but how/where can we view the result
 > to validate success?

Hi Chris,
Although that curl command succeeds the read the docs job still fails because the service expects the code to be anonymously cloneable, which it will be Monday.  We just wanted to get everything in place early.  Once the repos are public the job will work and the docs will be viewable here http://acumos-documentation.readthedocs.io/en/latest/

Patch-set: 3