Update patch set 3
Patch Set 3:

Another thought that I just realized. I think the whole purprose of the subproject configuration was to add "release" jobs for the subprojects right?

If so we probably don't need verify / merge jobs for the other levels either. because the top level project already handles verify and merge for all of the jobs. The release job is purely to allow the subprojects to produce separate releases.

Patch-set: 3