Update patch set 3
Patch Set 3:

> Owners of this repo?  That's *you* the LF rel-eng team.  This is
 > not a community managed repo, not yet.

Chris, while it's not a completely community managed repo, we do have very high expectations that the committers of a given project will review changes that affect _their_ project when it comes to the CI infrastructure.

We have not been enforcing this so strictly with newly formed projects, but as the project goes on, even if it never gets a community committer, we will stop merging patches unless the change was submitted by a committer of the affected project or at least given a +1 by it.

Patch-set: 3
Reviewer: Gerrit User 1 <1@655300a2-7a56-4504-9153-881edd907bea>
Label: Verified=0