Update patch set 1
Patch Set 1: Code-Review-1

Thank you for that explanation Manoop, it's much more clear than the e-mail was.

With that understood, I would still like to see a change to this: since it is a subproject of databroker, the path should be jjb/databroker/databroker-dataset.yaml, rather than creating its own jjb subdirectory.

Patch-set: 1
Label: Code-Review=-1