Update patch set 2
Patch Set 2:

> What is broken about the system-update role?

It looks to me like it's trying to call a local copy, but at some point the roles were moved to the cloud and this file wasn't updated. I assume this could be fixed in the same way that protobuf-install was, just by adding "lfit." to the beginning, but system-update is included in baseline that is imported at the top, so it seems to be redundant in these files.

Patch-set: 2
Reviewer: Gerrit User 223 <223@655300a2-7a56-4504-9153-881edd907bea>
Label: Code-Review=0