Update patch set 2
Patch Set 2:

I thought of adding this variable to jjb/defaults.yaml but the test in the sandbox didn't seemed to pick the change. 

Only after adding the variable in the job definition block, looks like it is reflecting properly. 

On the other side, having this variable for each job will allow flexibility on adding 'acumos-' prefix, or any other '<name>-' prefix or not adding it at all individually.

Patch-set: 2