Update patch set 1
Patch Set 1:

> Thanh this didn't work.  Before I flail around some more, do you
 > have any suggestions for telling Maven to build in a subdirectory?

Sorry I confused mvn-opts and mvn-params again. So if you change this to mvn-params it should work.

mvn-opts sets "MAVEN_OPTS" environment variable which is passed to Java as java-opts, this is not for setting Maven parameters. Instead mvn-params will pass it to the mvn cli as you'd expect.

Patch-set: 1