Update patch set 1
Patch Set 1:

The javadoc job is not current compatible with building a project that is not a root project. This is due to the line:

 mv "$WORKSPACE/target/site/apidocs" "$JAVADOC_DIR"

Which assumes that the apidocs can be found in the root of the workspace, which is not the case if building a sub-directory. I think the the short term disabling the javadoc job so that it doesn't keep failing verifies is the right thing.

Then work on improving the javadoc job to support alternative javadoc paths.

Eric, can you look into this from a global-jjb perspective?

Patch-set: 1