Update patch set 2
Patch Set 2:

Eric I see a possible problem here - I can limit these jobs to commits in this repo, but the parent jobs will also run when a commit is merged here, so releasing the artifacts may be tricky because the release repo will have too much.  Is there a negative gerrit trigger, ignore commits in a subdirectory?

Patch-set: 2