Update patch set 2
Patch Set 2:

I noticed today that the Global-JJB lf-maven-jobs macros and templates, as well as my variation in Acumos, all omit the gerrit_trigger_file_paths for the STAGE jobs.  This patch fixes the Acumos variation.  If you agree that it's appropriate to have trigger file paths on stage jobs, I will submit a similar patch for Global-JJB.

Patch-set: 2