Update patch set 1
Patch Set 1:

So (not sure if Acumos does things differently) but in case of other projects I have worked in like ONAP, HYP and O-RAN, the tech teams would identify which jobs they need for getting their code integrated in their repos. For example, they would definitely choose to have a verify and merge jobs to get their changes in the code. Most times, they will also add a CSIT if it exists or Sonar jobs to scan their tests. Release jobs are also available for those teams that would like to build their code and produce artifacts in Nexus 2 or Nexus 3. It all depends on what the tech team wants and needs to do.

Patch-set: 1
Label: Verified=0