Update patch set 1
Patch Set 1: Code-Review-1

Unless this is for a new project, the name of the repo and the project itself is license-manager, not license-management. Please change all instances of "license-management" in the commit message and license-manager.yaml.

Additionally, there is currently no content in the license-manager repo. I'd like to see at least an initial commit before merging these changes.

Patch-set: 1
Label: Code-Review=-1