Update patch set 8
Patch Set 8:

Eric I have been testing in the sandbox. For example https://jenkins.acumos.org/sandbox/job/acumos-c-client-cmake-verify-master/3/  The build fails because of this step in cmake-build.sh:

    tar -cJvf "$WORKSPACE/dist/$project.tar.xz" -C "$INSTALL_PREFIX" .

Obviously the wrapper expects to project to create a dist/{project}.tar.xz file.  What must be in that file?  What is the right cmake or make incantation to use here?

Patch-set: 8