Update patch set 8
Patch Set 8: Code-Review-1

(1 comment)

It looks like the solution is to make INSTALL_PREFIX=$BUILD_DIR (the default is INSTALL_PREFIX=$BUILD_DIR/output). I'm not sure if this is an error in the script or the default INSTALL_PREFIX, or if you guys are doing something different from the author of the cmake jobs (this is my first time working with them). But I do know that changing that parameter made the job work. https://jenkins.acumos.org/sandbox/job/acumos-c-client-cmake-verify-master/5/console

Patch-set: 8
Label: Code-Review=-1