Update patch set 1
Patch Set 1:

Please post the URL of the job that failed.

This means that the LF RelEng team failed to provide the maven settings file for this repository.  It's a manual step, apparently, that has to be done for each repo.  At this point I feel that missing settings files are the bane of my LF existence :/

Patch-set: 1