Update patch set 1
Patch Set 1:

Java 11 and Docker work perfectly in this Ubuntu 18.04 builder image, I tested in the sanbox.    I'm not waiting for repair of the CentOS7 image. Please +1 then I'll add LF team for review and merge. 

Andrew please submit a similar change for fed-gw and I'll add +1 and add LF team for review and merge.

Patch-set: 1