Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> I don't understand "avx" in the label. This image has docker, and previous images with that package used that word in its label.  Would you please consider renaming?

avx supercedes that. Most images are either labeled "builder" or "docker", but in this case, we're putting that image onto avx hardware, so it carries the avx label. There are currently no Ubuntu builder or docker nodes, and they won't be added unless they're needed. This node type was specifically requested for acumos-c-client.

Patch-set: 1