Update patch set 1
Patch Set 1:

It's our own fault that we're using both centos and ubuntu images. Maybe this big improvement in the centos images is a chance to drop use of the ubuntu images??  Are you able to report the differences in the images so we don't have to just guess & try?

Patch-set: 1