Update patch set 5
Patch Set 5: Code-Review-1

Ok so now we're back to about where we started.  I am trying to figure out if the jobs you have specified here are appropriate for the code in the repo.  But at first glance they do not appear to be appropriate.

The protobuf-common sub-project looks like a Java library.  It does not look like a docker project.  I see no Dockerfile, and the Maven pom.xml file does not direct construction of a docker image. Please explain?

The spark-modelrunner does not have a Dockerfile, but the Maven pom file has directions for building a docker image. Is the Dockerfile missing?

Patch-set: 5
Label: Code-Review=-1