Update patch set 1
Patch Set 1:

My change to correct broken link in acumos-r-client didn't start so I thought that is was due to the lack of jjb in acumos-r-client. So that is why I created this change, but finally the jenkins jobs started on acumos-r-client, so I will abandon this change. Sorry for the disturbance

Patch-set: 1
Reviewer: Gerrit User 212 <212@655300a2-7a56-4504-9153-881edd907bea>
Label: Code-Review=0