Update patch set 3
Patch Set 3: Code-Review-1

DW

This is supposed to bring in the version of release-job.sh that pins python package idna.  But I see other release jobs working fine WITHOUT the pin, for example:

https://jenkins.acumos.org/job/portal-marketplace-release-merge/4/console

What's going on? Maybe the Python community reacted to the change in idna, and global-jjb does not have to add a pin?

Patch-set: 3