Update patch set 1
Patch Set 1:

so this will strip the -SNAPSHOT suffix on every verify build and every merge build.  Why add a suffix in source control, only to take it away again immediately in the build job?  I am not a fan of this change.  Instead my advice is to drop usage of this Java convention from your project and never look back.

Patch-set: 1
CC: Gerrit User 4 <4@655300a2-7a56-4504-9153-881edd907bea>