Migrate write-good hook to pre-commit
This migrates write-good hook over to pre-commit which is the final
hook needed to remove our coala dependency.

Also resolve write-good suggestions on files except jenkins.rst (it is
currently too large a task for my time so place an exception there until
we have cycles to take care of it).

Ref: https://jira.linuxfoundation.org/browse/RELENG-2642
Change-Id: I9ce75330d6fc4c17aebe14d44aad372a709ab4a0
Signed-off-by: Anil Belur <abelur@linuxfoundation.org>